Move CITIES to separate folder and rename city prop in filters
